requirements.txt
django==2.1
django-cors-headers==2.4.0
djangorestframework==3.9.0
requests
jwt
